feat(core): city-wide-unique birth-tagged Entity (CW5c)
Entity becomes a packed u64: birth region in the high 32 bits, region-local
counter in the low 32 bits. World::spawn mints Entity::new(self.region_id,
next_entity), so ids are unique across the whole city with no shared allocator,
and the region tag is fixed at birth (survives relocation into another region's
World — the Model B foundation). Added region()/local() accessors + a pack
round-trip test.

Region 0 keeps packed == local, so bare-World tests and existing Entity(n)
literals are unchanged; within one World all entities share the region bits, so
Ord/sort order stays the local-counter order. Kept a numeric newtype so it still
serializes as a JSON map key. Save format changes (ids carry region) — old-save
compatibility intentionally dropped per the user's call; no consumer reads the
region bits yet.

Fallout: economy sentinels u32::MAX->u64::MAX; imported_job_slots and the
availability-hint spare_job_slot_ids use entity.local() (opaque per-region u32);
test-only DerivedSnapshot tuples widened to u64. Patch CW5c of
docs/city-wide-citizen-references-plan.md.

+++ src/core/simulation.rs
@@ -488,8 +488,8 @@ mod tests {
     struct DerivedSnapshot {
         stats: CityStats,
         local_effects: LocalEffectsMap,
-        morale_targets: Vec<(u32, i32)>,
-        assignments: Vec<(u32, Option<WorkplaceAssignment>)>,
+        morale_targets: Vec<(u64, i32)>,
+        assignments: Vec<(u64, Option<WorkplaceAssignment>)>,
     }
 
     impl DerivedSnapshot {
